requirements.txt
django-markdownify
Pillow